finish BigDecimal

+++ src/ru/nemesh/VisaGoldCard.java
@@ -1,8 +1,11 @@
 package ru.nemesh;
 
+import java.math.BigDecimal;
+
 public class VisaGoldCard extends Vcard {
 
-    public VisaGoldCard(String id, String cardHolder, double discount) {
+    public VisaGoldCard(String id, String cardHolder, int discount) {
         super(id, cardHolder, discount);
     }
+
 }
